Moving the method to detection utils and adding comments.
