+++ torchvision/models/detection/faster_rcnn.py
@@ -7,7 +7,7 @@ import torch.nn.functional as F
 from torchvision.ops import misc as misc_nn_ops
 from torchvision.ops import MultiScaleRoIAlign
 
-from .._utils import overwrite_eps
+from ._utils import overwrite_eps
 from ..utils import load_state_dict_from_url
 
 from .anchor_utils import AnchorGenerator
