+++ torchvision/models/detection/keypoint_rcnn.py
@@ -3,7 +3,7 @@ from torch import nn
 
 from torchvision.ops import MultiScaleRoIAlign
 
-from .._utils import overwrite_eps
+from ._utils import overwrite_eps
 from ..utils import load_state_dict_from_url
 
 from .faster_rcnn import FasterRCNN
